Moved routing and providers to the App component
